adds vis test for logical meshes, confirmed that vis-ing cell-based quantities on logical mesh still works
