+++ src/mesh/mesh_logical/CMakeLists.txt
@@ -79,7 +79,7 @@ if (BUILD_TESTS)
     set(mesh_logical_test_link_libs mesh_logical mesh mesh_mstk geometry ${UnitTest_LIBRARIES} ${HDF5_LIBRARIES})
 
     # Test: init
-    set(mesh_logical_test_src_files test/Main.cc test/demo_mesh.cc test/mesh_logical_geometry.cc)
+    set(mesh_logical_test_src_files test/Main.cc test/mesh_logical_geometry.cc)
 
     add_amanzi_test(mesh_logical_geometry mesh_logical_geometry
       KIND unit
